backend/app/parsers/docx_parser.py
from pathlib import Path

from docx import Document


def parse_docx(file_path: str | Path) -> str:
    doc = Document(str(file_path))
    paragraphs = [p.text for p in doc.paragraphs if p.text.strip()]
    return "\n\n".join(paragraphs)
