backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
httpx==0.28.1
sse-starlette==2.2.1
python-multipart==0.0.20

# LangChain
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langchain-qdrant==0.2.0

# Embedding
sentence-transformers==3.3.1

# Vector DB
qdrant-client==1.12.1

# Document parsing
PyMuPDF==1.25.3
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
